backend/admin/requirements.txt
appdirs==1.4.4
asgiref==3.5.2
attrs==21.4.0
backoff==2.1.2
cfgv==3.3.1
click==8.1.3
distlib==0.3.4
Django==4.0.4
django-cors-headers==3.13.0
django-debug-toolbar==3.4.0
django-split-settings==1.1.0
djangorestframework==3.13.1
enforce-typing==1.0.0.post1
filelock==3.7.1
flake8==4.0.0
identify==2.5.1
iniconfig==1.1.1
install==1.3.5
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
oitnb==0.2.2
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.19.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pydantic==1.9.1
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.2.0
uWSGI==2.0.20
virtualenv==20.14.1